# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VelesDB API
  description: High-performance vector database for AI applications. Supports semantic search, HNSW indexing, and multiple distance metrics. Authentication is optional — when API keys are configured via VELESDB_API_KEYS, all endpoints except /health and /ready require a valid Bearer token.
  contact:
    name: VelesDB Team
    url: https://github.com/cyberlife-coder/VelesDB
  license:
    name: VelesDB Core License 1.0
    url: https://github.com/cyberlife-coder/VelesDB/blob/main/LICENSE
  version: 6.0.0
servers:
- url: /
  description: Local server
paths:
...
  /collections/{name}/search/multi:
    post:
      tags:
      - search
      summary: Multi-query search with fusion strategies.
      operationId: multi_query_search
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiQuerySearchRequest'
        required: true
      responses:
        '200':
          description: Multi-query search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /collections/{name}/search/multi/ids:
    post:
      tags:
      - search
      summary: Multi-query fusion search returning only ids and scores (no payloads).
      description: |-
        Faster than `/search/multi` when payloads are not needed: the core
        `multi_query_search_ids` kernel skips payload hydration. Metadata filters
        are not supported here — use `/search/multi` for filtered fusion.
      operationId: multi_query_search_ids
      parameters:
      - name: name
        in: path
        description: Collection name
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiQuerySearchRequest'
        required: true
      responses:
        '200':
          description: Multi-query ids-only results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchIdsResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Collection not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 'A body the server cannot deserialise: a field of the wrong type, or JSON that does not match the request schema. Axum''s extractor answers before the handler runs, so the payload is plain text, not an ErrorResponse.'
          content:
            text/plain:
              schema:
                type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Error response.
      required:
      - error
      properties:
        code:
          type:
          - string
          - 'null'
          description: Structured error code (e.g., `"VELES-004"` for dimension mismatch).
        error:
          type: string
          description: Error message.
...
    IdScoreResult:
      type: object
      description: A single ID+score result from IDs-only search.
      required:
      - id
      - score
      properties:
        id:
          type: string
          description: Point ID.
        score:
          type: number
          format: float
          description: Similarity score.
...
    MultiQuerySearchRequest:
      type: object
      description: Request for multi-query vector search with fusion.
      required:
      - vectors
      properties:
        avg_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for average score component.'
          example: 0.5
        dense_weight:
          type: number
          format: float
          description: 'Relative score fusion: weight for the dense (vector) branch.'
          example: 0.5
        filter:
          type: object
          description: Optional metadata filter.
          example:
            condition:
              field: category
              type: eq
              value: tech
        hit_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for hit count component.'
          example: 0.2
        max_weight:
          type: number
          format: float
          description: 'Weighted fusion: weight for max score component.'
          example: 0.3
        rrf_k:
          type: integer
          format: int32
          description: RRF k parameter (only used when strategy = "rrf").
          example: 60
          minimum: 0
        sparse_weight:
          type: number
          format: float
          description: 'Relative score fusion: weight for the sparse branch.'
          example: 0.5
        strategy:
          type: string
          description: 'Fusion strategy: "average", "maximum", "rrf", "weighted", "`relative_score`".'
          example: rrf
        top_k:
          type: integer
          description: Number of results to return.
          example: 10
          minimum: 0
        vectors:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
          description: List of query vectors.
...
    SearchIdsResponse:
      type: object
      description: Response from IDs-only search (no payload hydration).
      required:
      - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/IdScoreResult'
          description: Search results with IDs and scores only.
...
    SearchResponse:
      type: object
      description: Response from vector search.
      required:
      - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultResponse'
          description: Search results.
    SearchResultResponse:
      type: object
      description: A single search result.
      required:
      - id
      - score
      properties:
        id:
          type: string
          description: Point ID.
        payload:
          description: Point payload.
        score:
          type: number
          format: float
          description: Similarity score.
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...